+++ zcash_client_sqlite/src/testing/pool.rs
@@ -305,8 +305,22 @@ pub(crate) fn data_db_truncation<T: ShieldedPoolTester>() {
     )
 }
 
-pub(crate) fn truncate_to_chain_state<T: ShieldedPoolTester>() {
-    zcash_client_backend::data_api::testing::pool::truncate_to_chain_state::<T, _>(
+pub(crate) fn truncate_to_chain_state_deep<T: ShieldedPoolTester>() {
+    zcash_client_backend::data_api::testing::pool::truncate_to_chain_state_deep::<T, _>(
+        TestDbFactory::default(),
+        BlockCache::new(),
+    )
+}
+
+pub(crate) fn truncate_to_chain_state_shallow<T: ShieldedPoolTester>() {
+    zcash_client_backend::data_api::testing::pool::truncate_to_chain_state_shallow::<T, _>(
+        TestDbFactory::default(),
+        BlockCache::new(),
+    )
+}
+
+pub(crate) fn truncate_after_non_contiguous_scan<T: ShieldedPoolTester>() {
+    zcash_client_backend::data_api::testing::pool::truncate_after_non_contiguous_scan::<T, _>(
         TestDbFactory::default(),
         BlockCache::new(),
     )